+++ data/payload-refresh/md-transition-runtime-manual/installer/install.sh
@@ -1,0 +1,2 @@
+#!/bin/ash
+exec /usr/sbin/nokia-ubi-installer "$@"